Created the likes and comments features for the games
