+++ Backend/controllers/gameController.js
@@ -34,12 +34,7 @@ const getAllGames = async (req, res) => {
 
 const getGameById = async (req, res) => {
   try {
-    const gameId = Number(req.params.id);
-    if (isNaN(gameId))
-      return res
-        .status(400)
-        .json({ message: "Invalid game ID - must be a number" });
-
+    const gameId = req.params.id; // Keep as string
     const game = await Game.findById(gameId);
     if (!game) return res.status(404).json({ message: "Game not found" });
 
@@ -52,18 +47,22 @@ const getGameById = async (req, res) => {
 
 const likeGame = async (req, res) => {
   try {
-    const gameId = Number(req.params.id);
-    if (isNaN(gameId))
-      return res.status(400).json({ message: "Invalid game ID" });
-
+    const gameId = req.params.id;
     const game = await Game.findById(gameId);
     if (!game) return res.status(404).json({ message: "Game not found" });
 
-    if (!game.likes.includes(req.user._id)) {
+    const userId = req.user._id.toString();
+    const hasLiked = game.likes.some((id) => id.toString() === userId);
+
+    if (hasLiked) {
+      game.likes = game.likes.filter((id) => id.toString() !== userId);
+      await game.save();
+      res.json({ message: "Like removed", likes: game.likes });
+    } else {
       game.likes.push(req.user._id);
       await game.save();
+      res.json({ message: "Game liked", likes: game.likes });
     }
-    res.json({ message: "Game liked", likes: game.likes });
   } catch (err) {
     console.error("Error in likeGame:", err.message);
     res.status(500).json({ message: err.message });
@@ -72,10 +71,8 @@ const likeGame = async (req, res) => {
 
 const addComment = async (req, res) => {
   try {
-    const gameId = Number(req.params.id);
+    const gameId = req.params.id;
     const { commentText } = req.body;
-    if (isNaN(gameId))
-      return res.status(400).json({ message: "Invalid game ID" });
     if (!commentText)
       return res.status(400).json({ message: "Comment text is required" });
 
@@ -86,7 +83,7 @@ const addComment = async (req, res) => {
     if (!game) return res.status(404).json({ message: "Game not found" });
 
     game.comments.push({
-      email_id: user.email_id, // Updated
+      email_id: user.email_id,
       displayName: user.display_name,
       commentText,
       time: new Date(),
@@ -102,22 +99,26 @@ const addComment = async (req, res) => {
 
 const likeComment = async (req, res) => {
   try {
-    const gameId = Number(req.params.id);
+    const gameId = req.params.id;
     const { commentId } = req.params;
-    if (isNaN(gameId))
-      return res.status(400).json({ message: "Invalid game ID" });
-
     const game = await Game.findById(gameId);
     if (!game) return res.status(404).json({ message: "Game not found" });
 
     const comment = game.comments.id(commentId);
     if (!comment) return res.status(404).json({ message: "Comment not found" });
 
-    if (!comment.likes.includes(req.user._id)) {
+    const userId = req.user._id.toString();
+    const hasLiked = comment.likes.some((id) => id.toString() === userId);
+
+    if (hasLiked) {
+      comment.likes = comment.likes.filter((id) => id.toString() !== userId);
+      await game.save();
+      res.json({ message: "Like removed from comment", likes: comment.likes });
+    } else {
       comment.likes.push(req.user._id);
       await game.save();
+      res.json({ message: "Comment liked", likes: comment.likes });
     }
-    res.json({ message: "Comment liked", likes: comment.likes });
   } catch (err) {
     console.error("Error in likeComment:", err.message);
     res.status(500).json({ message: err.message });
@@ -126,11 +127,9 @@ const likeComment = async (req, res) => {
 
 const replyToComment = async (req, res) => {
   try {
-    const gameId = Number(req.params.id);
+    const gameId = req.params.id;
     const { commentId } = req.params;
     const { commentText } = req.body;
-    if (isNaN(gameId))
-      return res.status(400).json({ message: "Invalid game ID" });
     if (!commentText)
       return res.status(400).json({ message: "Reply text is required" });
 
@@ -145,7 +144,7 @@ const replyToComment = async (req, res) => {
 
     comment.replies = comment.replies || [];
     comment.replies.push({
-      email_id: user.email_id, // Updated
+      email_id: user.email_id,
       displayName: user.display_name,
       commentText,
       time: new Date(),
@@ -161,11 +160,8 @@ const replyToComment = async (req, res) => {
 
 const deleteComment = async (req, res) => {
   try {
-    const gameId = Number(req.params.id);
+    const gameId = req.params.id;
     const { commentId } = req.params;
-    if (isNaN(gameId))
-      return res.status(400).json({ message: "Invalid game ID" });
-
     const game = await Game.findById(gameId);
     if (!game) return res.status(404).json({ message: "Game not found" });
 
@@ -173,7 +169,6 @@ const deleteComment = async (req, res) => {
     if (!comment) return res.status(404).json({ message: "Comment not found" });
 
     if (comment.email_id !== req.user.email_id) {
-      // Updated
       return res
         .status(403)
         .json({ message: "You can only delete your own comments" });
@@ -190,11 +185,8 @@ const deleteComment = async (req, res) => {
 
 const deleteReply = async (req, res) => {
   try {
-    const gameId = Number(req.params.id);
+    const gameId = req.params.id;
     const { commentId, replyId } = req.params;
-    if (isNaN(gameId))
-      return res.status(400).json({ message: "Invalid game ID" });
-
     const game = await Game.findById(gameId);
     if (!game) return res.status(404).json({ message: "Game not found" });
 
@@ -205,7 +197,6 @@ const deleteReply = async (req, res) => {
     if (!reply) return res.status(404).json({ message: "Reply not found" });
 
     if (reply.email_id !== req.user.email_id) {
-      // Updated
       return res
         .status(403)
         .json({ message: "You can only delete your own replies" });
@@ -220,30 +211,36 @@ const deleteReply = async (req, res) => {
   }
 };
 
-const likeReply = async (req, res) => {
+const likeReply = async (req, reply) => {
   try {
-    const gameId = Number(req.params.id);
+    const gameId = req.params.id;
     const { commentId, replyId } = req.params;
-    if (isNaN(gameId))
-      return res.status(400).json({ message: "Invalid game ID" });
-
     const game = await Game.findById(gameId);
-    if (!game) return res.status(404).json({ message: "Game not found" });
+    if (!game) return reply.status(404).json({ message: "Game not found" });
 
     const comment = game.comments.id(commentId);
-    if (!comment) return res.status(404).json({ message: "Comment not found" });
+    if (!comment)
+      return reply.status(404).json({ message: "Comment not found" });
 
-    const reply = comment.replies.id(replyId);
-    if (!reply) return res.status(404).json({ message: "Reply not found" });
+    const replyDoc = comment.replies.id(replyId);
+    if (!replyDoc)
+      return reply.status(404).json({ message: "Reply not found" });
 
-    if (!reply.likes.includes(req.user._id)) {
-      reply.likes.push(req.user._id);
+    const userId = req.user._id.toString();
+    const hasLiked = replyDoc.likes.some((id) => id.toString() === userId);
+
+    if (hasLiked) {
+      replyDoc.likes = replyDoc.likes.filter((id) => id.toString() !== userId);
       await game.save();
+      reply.json({ message: "Like removed from reply", likes: replyDoc.likes });
+    } else {
+      replyDoc.likes.push(req.user._id);
+      await game.save();
+      reply.json({ message: "Reply liked", likes: replyDoc.likes });
     }
-    res.json({ message: "Reply liked", likes: reply.likes });
   } catch (err) {
     console.error("Error in likeReply:", err.message);
-    res.status(500).json({ message: err.message });
+    reply.status(500).json({ message: err.message });
   }
 };
